added more programs

+++ Lab 1 (4th Sep 23)/days_to_years.cpp
@@ -9,7 +9,7 @@ int main(){
     scanf("%d", &days);
 
     //Converting days into Years
-    years = days / 365;
+    years = (double)days/ 365;
 
     //Displaying Results
     printf("%d days equals to %lf in years. \n",days, years);
